moved functions that serves html to 'html.go'
