+++ controllers/html.go
@@ -1,0 +1,48 @@
+package controllers
+
+import (
+	"net/http"
+	"net/url"
+
+	"github.com/gin-gonic/gin"
+)
+
+// Serves "login" page
+func LoginGET() gin.HandlerFunc {
+	return func(ctx *gin.Context) {
+		ctx.HTML(
+			http.StatusOK, "login.html", nil,
+		)
+	}
+}
+
+// Redirects the user to the login page
+func LoginRedirect() gin.HandlerFunc {
+	return func(ctx *gin.Context) {
+		location := url.URL{Path: "/login"}
+		ctx.Redirect(http.StatusFound, location.RequestURI())
+	}
+}
+
+// Serves "cadastro.html" page
+func SignupGET() gin.HandlerFunc {
+	return func(ctx *gin.Context) {
+		ctx.HTML(
+			http.StatusOK, "cadastro.html", nil,
+		)
+	}
+}
+
+// Serves "aluno.html" page
+func GetAlunoPage() gin.HandlerFunc {
+	return func(ctx *gin.Context) {
+		ctx.HTML(http.StatusOK, "aluno.html", nil)
+	}
+}
+
+// Serves "alunoinfo.html" page
+func GetUserInfoPage() gin.HandlerFunc {
+	return func(ctx *gin.Context) {
+		ctx.HTML(http.StatusOK, "alunoinfo.html", nil)
+	}
+}
