+++ controllers/signup.go
@@ -8,15 +8,6 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-// At the moment, this function just serves the html file
-func SignupGET() gin.HandlerFunc {
-	return func(ctx *gin.Context) {
-		ctx.HTML(
-			http.StatusOK, "cadastro.html", nil,
-		)
-	}
-}
-
 // Gets the values from the form, creates a user and inserts it in the database
 func SignupPOST() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
